% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.sequences.R
\name{align.sequences}
\alias{align.sequences}
\title{align.sequences}
\usage{
align.sequences(
  file = NULL,
  sequences = NULL,
  substitutionMatrix = "BLOSUM50",
  pid.type = "PID1",
  paperWidth = 7.5,
  shadingMode = "similar",
  consensusColors = "ColdHot",
  shadingColors = "reds",
  showConsensus = "none",
  showLegend = TRUE
)
}
\arguments{
\item{file}{Path to the fasta file.}

\item{substitutionMatrix}{Argument as in Biostrings::pairwiseAlignment}
}
\value{
A list containing:
\item{alignment}{Alignment between two sequences.}
\item{identity}{Identity percentage.}
}
\description{
Align sequences and export a nice alignment figure
}
\examples{
align.sequences("toAlign.fasta")
}
