% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.DIAPELSA.R
\name{analyze.DIAPELSA}
\alias{analyze.DIAPELSA}
\title{Differential Stability Analysis by PELSA from DIA-NN Output}
\usage{
analyze.DIAPELSA(
  file,
  extract.after = "^",
  extract.before = "_",
  ctrl.name = "Ctrl",
  FC.cutoff = c(1.2, 1.2),
  p.cutoff = c(0.01, 0.05),
  p.adjust = "BH"
)
}
\arguments{
\item{file}{Path to the DIA-NN \code{.parquet} report file for DIA data or to PD-exported txt file for DDA.}

\item{extract.after}{Regex pattern to extract group names from sample names (after a specific pattern).}

\item{extract.before}{Regex pattern to extract group names from sample names (before a specific pattern).}

\item{ctrl.name}{Name of the control group (e.g., "Ctrl").}

\item{FC.cutoff}{Fold change cutoff for determining significance for hits (first value) and candidates (second value) (default is c(1.2,1.2)).}

\item{p.cutoff}{Adjusted p-value cutoff for determining significance for hits (first value) and candidates (second value) (default is c(0.01,0.05)).}

\item{p.adjust}{p-value adjustment method in limma::topTable}
}
\value{
A named list with the following elements:
\describe{
\item{all_results}{A data frame of differential analysis results.}
\item{volcano_facet}{A \code{ggplot} object showing faceted volcano plots.}
\item{volcano_list}{A named list of individual volcano plots per contrast.}
\item{MD_plot_faceted}{A \code{ggplot} object showing faceted MD plots.}
}
}
\description{
This function performs differential stability analysis on protein abundance data from a DIA-NN \code{.parquet} file.
It uses the \code{limma} package to compute moderated statistics, generates volcano plots, and MD plots.
}
\details{
The implementation is heavily inspired by code from the limma-for-proteomics
GitHub repository: \url{https://github.com/41ison/limma-for-proteomics}
}
