% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.statistics.R
\name{B.statistics}
\alias{B.statistics}
\title{B.statistics}
\usage{
B.statistics(
  data = NULL,
  control = NA,
  fdr_hit_threshold = 0.05,
  fdr_candidate_threshold = 0.25,
  fc_hit_threshold = 1,
  fc_candidate_threshold = 0.5
)
}
\arguments{
\item{data}{A data frame with names id, condition, replicate and value.}

\item{control}{Name of the control condition (as in the data).}

\item{fdr_hit_threshold}{FDR threshold for hits.}

\item{fdr_candidate_threshold}{FDR threshold for candidates.}

\item{fc_hit_threshold}{FC threshold for hits.}

\item{fc_candidate_threshold}{FC threshold for candidates.}
}
\value{
A list containing:
\item{data}{Resulting data table.}
\item{hitcounts}{Counts of hits.}
\item{limma_vs_FDR}{Comparison of limma vs FDR p-value estimations.}
\item{volcano}{Volcano plot.}
\item{volcano.separate}{Separate volcano plots for each comparison.}
\item{MA}{MA plot.}
}
\description{
Calculate Bayesian statistics.
}
\examples{
B.statistics(data,"control")
}
