% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.heatmap.R
\name{clustered.heatmap}
\alias{clustered.heatmap}
\title{clustered.heatmap}
\usage{
clustered.heatmap(
  data = data.frame(),
  cl.group = "all",
  plot.reverse = TRUE,
  plot.ddg.space = 0.05
)
}
\arguments{
\item{data}{A data frame with columns "id","group","condition" and "value".}

\item{cl.group}{Based on which group should the clustering be done? One of the values in group column or "all".}

\item{plot.reverse}{Boolean: Should the plot be turn around? Try out to see what happens.
plot.ddg.space Space between the plot and the dendrogram. Try iterratively.}
}
\value{
A list containing:
\item{plot}{Calculated IC50.}
\item{clustering}{Clustering object by hclust}
\item{dendrogram}{Dendrogram object}
\item{heatmap}{Drawn heatmap.}
\item{legend}{Plot legend.}
}
\description{
Draws a clustered heatmap.
}
\examples{
calcIC50(data)
}
