% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euler_plot.R
\name{euler_plot}
\alias{euler_plot}
\title{euler_plot}
\usage{
euler_plot(
  sets = NULL,
  set.shape = "circle",
  plot.name = "",
  plot.label.size = 0,
  quantities.type = "counts",
  quantities.font.type = "plain",
  quantities.font.size = 1,
  color.scale = "rainbow",
  custom.scale = NULL,
  plot.euler = TRUE
)
}
\arguments{
\item{sets}{A data frame with two columns: 'id' (set identifiers) and 'condition'.}

\item{set.shape}{A character string specifying the shape of the Euler diagram. Can be either "circle" or "ellipse".}

\item{plot.name}{A character string representing the title of the plot.}

\item{plot.label.size}{A numeric value indicating the font size of the labels in the plot.}

\item{quantities.type}{A character string indicating the type of quantities to display, typically "counts" or other quantification types.}

\item{quantities.font.type}{A character string specifying the font style for quantities, such as "plain", "bold", or "italic".}

\item{quantities.font.size}{A numeric value specifying the font size for the quantities in the plot.}

\item{color.scale}{A character string specifying the color scale to use. Can be "rainbow" or "distinct".}

\item{custom.scale}{A vector of colors for custom color scaling.}

\item{plot.euler}{A logical value indicating whether to generate the Euler plot. Default is TRUE.}
}
\value{
A list containing:
\item{counts}{A named list of the number of elements in the intersections of each combination of sets.}
\item{combinations}{A named list of the actual sets of elements for each combination of sets.}
\item{union_counts}{A named list of the number of elements in the union of each combination of sets.}
\item{unions}{A named list of the actual sets of elements in the union for each combination of sets.}
\item{euler_diagram}{A plot object containing the generated Euler diagram (or NULL if \code{plot.euler} is FALSE).}
\item{legend}{A grob (graphical object) representing the legend for the plot.}
\item{plot}{A ggplot object representing the generated Euler plot.}
}
\description{
This function generates an Euler diagram to visualize the relationships between multiple sets.
}
\examples{
euler_plot(sets)
}
