% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cos.R
\name{fit_cos}
\alias{fit_cos}
\title{fit_cos}
\usage{
fit_cos(data, x_col, y_col, period = 5)
}
\arguments{
\item{data}{A data frame with columns for x and y. The data must be z-scaled!}

\item{x_col}{Unquoted name of the x column.}

\item{y_col}{Unquoted name of the y column.}

\item{period}{Numeric value of period, e.g. 42 for 42h-long IDC of P. falciparum.}
}
\value{
A list containing:
\item{fit}{Fit object}
\item{coef}{Fitted coefficients.}
\item{R2}{R2 goodness of fit value.}
}
\description{
Fits cosine function into data.
}
\examples{
fit_cos(periodic_data,condition,z.log2FPKM,42)
}
