% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEA.R
\name{GEA}
\alias{GEA}
\title{GEA}
\usage{
GEA(
  pathways = data.frame(),
  genes = NULL,
  minsize = 5,
  p.cutoff = 0.05,
  p.adj.method = "BH",
  plot.padj.cutoff = 0.05,
  plot.stat = "HP"
)
}
\arguments{
\item{pathways}{Data frame with columns 'id','pathway.type' and 'pathway.name', or at least a
data frame with three columns and this content.}

\item{genes}{Vector with gene IDs, must have overlap with IDs in the pathways data frame.}

\item{minsize}{minimal size of gene set to test.}

\item{p.cutoff}{Numeric: p-value cutoff. Default is 0.05.}

\item{plot.padj.cutoff}{Numeric: Adjusted p-value cutoff for plotting. Default is 0.05.}

\item{plot.stat}{Character string: which p-value should be used for plotting? "HP" for hypergeometric
p-value and "BP" for binomial.}
}
\value{
A list with two elements:
\item{data}{Data with the enrichment result in a table form}
\item{plot}{Enrichment plot}
}
\description{
Gene  enrichment analysis
}
\examples{
gea.result <- GEA(pathways,genes)
}
