% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{GSEA}
\usage{
GSEA(
  pathways = data.frame(),
  ranks = data.frame(),
  types = "all",
  collapse = TRUE,
  minsize = 5,
  maxsize = 200,
  p.cutoff = 0.05,
  scoreType = "std"
)
}
\arguments{
\item{pathways}{Data frame with columns 'id','pathway.type' and 'pathway.name', or at least a
data frame with three columns and this content.}

\item{ranks}{Data fram with two columns 'id' and 'rank' or at least a data frame with two columns and this content.}

\item{types}{Vector: what types of enrichment to do? Choose from pathway.type of pathways or simply "all".}

\item{collapse}{Logical. Should the pathways be collapse to simplify the output? Default is TRUE.}

\item{minsize}{minimal size of gene set to test.}

\item{maxsize}{maximal size of gene set to test.}

\item{p.cutoff}{Numeric: p-value cutoff. Default is 0.05.}

\item{scoreType}{Same as in fgseaSimple.}
}
\value{
A list with four elements:
\item{enrichment.table}{Data with the full GSEA results.}
\item{enriched.table}{Data with the enriched pathways}
\item{plot}{GSEA distribution plots for each enriched term.}
\item{enrichment.plot}{Summarizing enrichment plot.}
}
\description{
Gene set enrichment analysis
}
\examples{
gsea.result <- GSEA(pathways,ranks)
}
