% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_files.R
\name{move_files}
\alias{move_files}
\title{move_files}
\usage{
move_files(from, to, except = character())
}
\arguments{
\item{from}{Relative or absolute path from which to copy.}

\item{to}{Relative or absolute path to which to copy, can be a new folder as well.}

\item{except}{Vector of names of files that should not be copied.}
}
\value{
Nothing. The function just moves files.
}
\description{
Move files into a new destination, with exceptions. Never moves folders.
}
\examples{
move_files(".","newfolder",except="mycode.R")
}
