% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.data.R
\name{prepare.data}
\alias{prepare.data}
\title{prepare.data}
\usage{
prepare.data(data = NULL, cols = NULL, no.cols = FALSE)
}
\arguments{
\item{data}{Data frame to check and prepare.}

\item{cols}{Column names expected in the data frame}

\item{no.cols}{Logical: if not all column names are found, then at least
if the number of columns matches, they will just be renamed.}
}
\value{
A list with three elements:
$data is a data frame with the prepared data
$message is the message to show to the user
$error says whether an error should be returned
}
\description{
Checks whether the columns expected from the data are there and filters on them.
}
