% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_plasmodium_fasta.R
\name{reformat_plasmodium_fasta}
\alias{reformat_plasmodium_fasta}
\title{Reformat Plasmodium FASTA Headers}
\usage{
reformat_plasmodium_fasta(input, output)
}
\arguments{
\item{input}{Character. Path to the input FASTA file to be reformatted.}

\item{output}{Character. Path where the reformatted FASTA file will be saved.}
}
\description{
This function reads a FASTA file with Plasmodium protein headers and reformats
the headers into a standardized UniProt-like format. It handles both normal and
reversed entries (with ">rev_" prefix).
}
\examples{
\dontrun{
# Reformat a Plasmodium FASTA file
reformat_plasmodium_fasta("raw_plasmodium.fasta", "formatted_plasmodium.fasta")
}
}
