% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{upset_plot}
\alias{upset_plot}
\title{upset_plot}
\usage{
upset_plot(sets = data.frame(), ...)
}
\arguments{
\item{sets}{Data frame with columns 'id' and 'condition'.}

\item{...}{Any arguments of UpSetR::upset().}
}
\value{
A list with two elements:
\item{data}{Enrichment results in table form.}
\item{plot}{Summarizing enrichment plot.}
$data with the enrichment result in a table form,
$plot for the enrichment plot.
}
\description{
Plots upset and Euler plots from gene sets.
}
\examples{
upset_plot(upsetdata)
}
